pyproject.toml
[project]
name = "deepclaw"
version = "0.1.0"
description = "Multi-platform AI agent gateway built on DeepAgents"
requires-python = ">=3.11,<4.0"
dependencies = [
    "deepagents",
    "deepagents-cli",
    "python-telegram-bot>=22.0,<23.0",
    "langgraph-checkpoint-sqlite>=3.0.0,<4.0.0",
    "croniter>=2.0,<4.0",
]

[project.optional-dependencies]
web = ["tavily-python>=0.7.21,<1.0.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
deepclaw = "deepclaw.cli:main"
deepclaw-tui = "deepagents_cli:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "PIE",  # flake8-pie
    "RET",  # flake8-return
    "LOG",  # flake8-logging
    "G",    # flake8-logging-format
    "T20",  # flake8-print
    "PLW",  # pylint warnings
]
ignore = [
    "E501",   # line too long (formatter handles this)
    "RET504", # unnecessary assignment before return
]

[tool.ruff.lint.isort]
known-first-party = ["deepclaw"]

[tool.uv.sources]
deepagents = { path = "../deepagents/libs/deepagents", editable = true }
deepagents-cli = { path = "../deepagents/libs/cli", editable = true }
